refactor: move loose src files into util/ and extension/ subdirectories
Logger and debounce moved to src/util/; cursorTracker and editListener
moved to src/extension/ where they belong. All import paths updated
across backend, pipeline, and extension layers.
