+++ elevate/src/util/Logger.ts
